Moved methods checkRound and checkLife from AbstractLevel to ModelImpl
