+++ src/main/java/it/unibo/game/app/api/Level.java
@@ -8,8 +8,7 @@ public interface Level {
     int increaseLife();
     int decreaseLife();
     boolean isAlive();
-    boolean checkRound();
     Round getRound();
     int getNumRoundPassed();
-    boolean checkLife();
+    void increaseRound();
 }
