+++ src/main/java/it/unibo/game/app/model/levels/AbstractLevel.java
@@ -4,7 +4,6 @@ import it.unibo.game.Pair;
 import it.unibo.game.app.api.Level;
 import it.unibo.game.app.api.Round;
 import it.unibo.game.app.model.SizeCalculation;
-import it.unibo.game.app.model.round.GameOver;
 
 
 /*Questa classe astratta dichiara variabili e definisce metodi in comune ai vari livelli */
@@ -12,7 +11,7 @@ import it.unibo.game.app.model.round.GameOver;
 public abstract class AbstractLevel implements Level {
 
     private final static int INITIAL_LIVES=3;
-    protected int lives = INITIAL_LIVES;
+    protected static int lives = INITIAL_LIVES;
 
     /*Definire delle costanti per ogni livello che definiscno queste variabili */
     protected int normalBricksFirstRound;
@@ -26,7 +25,6 @@ public abstract class AbstractLevel implements Level {
 
     protected Round currentRound;
     protected SizeCalculation sizeCalc;
-    private GameOver gameOver = new GameOver(currentRound);
 
     protected Pair<Double, Double> worldSize;
     protected int numRoundPassed = 0;
@@ -52,23 +50,7 @@ public abstract class AbstractLevel implements Level {
     public int getNumRoundPassed() {
         return this.numRoundPassed;
     }
-    public boolean checkRound() {
-        if(gameOver.isRoundFinished()) {
-            this.numRoundPassed++;
-            return true;
-        }else {
-            return false;
-        }
-    }
-    public boolean checkLife() {
-        if(this.gameOver.hasMissedBall()) {
-            this.decreaseLife();
-        }
-        if(this.isAlive()) {
-            return true;
-        }
-        else {
-            return false;
-        }                        
+    public void increaseRound() {
+        ++this.numRoundPassed;
     }
 }
